# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Order API",
        "description": "API for managing food orders",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category Controller"
                ],
                "summary": "Get all categories",
                "operationId": "d9bf255011b42d56bf306832dc3366c8",
                "responses": {
                    "200": {
                        "description": "Get all categories."
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Category Controller"
                ],
                "summary": "Get details of a category by ID",
                "operationId": "d809927cce155b8469de92d4c76f9288",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details."
                    },
                    "404": {
                        "description": "Category not found."
                    }
                }
            }
        },
        "/api/v1/categories/{category}/products": {
            "get": {
                "tags": [
                    "Category Controller"
                ],
                "summary": "Find products by category",
                "operationId": "ec52ec0d1956191887c6d77ab7eb4116",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Name of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products in the category."
                    },
                    "404": {
                        "description": "Category not found."
                    }
                }
            }
        },
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Order Controller"
                ],
                "summary": "Get all orders",
                "operationId": "2fd056a400bd274a5b2278e92a8ed010",
                "responses": {
                    "200": {
                        "description": "Get all orders."
                    }
                }
            },
            "post": {
                "tags": [
                    "Order Controller"
                ],
                "summary": "Create a new order",
                "operationId": "f92c2f898a00f407e1efa0fd23c93b24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "total_price",
                                    "status",
                                    "products"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string"
                                    },
                                    "table_id": {
                                        "type": "integer"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "tags": [
                    "Order Controller"
                ],
                "summary": "Get an order by ID",
                "operationId": "d9752171673795b587d678236ac33a2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details."
                    },
                    "404": {
                        "description": "Order not found."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order Controller"
                ],
                "summary": "Delete an order by ID",
                "operationId": "771df4bb7a25c0bd3287c2ba3ccbcff4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully."
                    },
                    "404": {
                        "description": "Order not found."
                    }
                }
            }
        },
        "/api/v1/orders/{orderId}/change-status": {
            "put": {
                "tags": [
                    "Order Controller"
                ],
                "summary": "Update status of an order",
                "operationId": "f2e27ed5197f8184f15cb94fe75b8282",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated successfully."
                    },
                    "404": {
                        "description": "Order not found."
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "Get all products",
                "operationId": "7ca3593d05653d957ae7be1ac648ca44",
                "responses": {
                    "200": {
                        "description": "Get all products."
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "Get a product by ID",
                "operationId": "1948ab0c6b14a722ad6d1665cd38e1f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details."
                    },
                    "404": {
                        "description": "Product not found."
                    }
                }
            }
        },
        "/api/v1/products/search": {
            "get": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "Search products by name",
                "operationId": "4b5c76b8d92c8bc2d86df4e3d2c7e58f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Keyword to search for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products matching the search."
                    }
                }
            }
        },
        "/api/v1/reviews": {
            "get": {
                "tags": [
                    "Review Controller"
                ],
                "summary": "Get all reviews",
                "operationId": "089acaacc3d898976a1a2793a732b604",
                "responses": {
                    "200": {
                        "description": "Get all reviews."
                    }
                }
            },
            "post": {
                "tags": [
                    "Review Controller"
                ],
                "summary": "Create a new review",
                "operationId": "050ff9620bfceb90b7a67100678b39a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer",
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "customer": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "int",
                                        "example": "4"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/api/v1/tables": {
            "get": {
                "tags": [
                    "Table Controller"
                ],
                "summary": "Get all tables",
                "operationId": "1b08f80199278b1e89a3b73113205fa6",
                "responses": {
                    "200": {
                        "description": "Get all tables."
                    }
                }
            }
        },
        "/api/v1/tables/{id}": {
            "get": {
                "tags": [
                    "Table Controller"
                ],
                "summary": "Get details of a table by ID",
                "operationId": "58fb0fb2b4167651835dea95e4209ac0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the table",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table details."
                    },
                    "404": {
                        "description": "Table not found."
                    }
                }
            }
        },
        "/api/v1/tables/{table_id}/orders": {
            "get": {
                "tags": [
                    "Table Controller"
                ],
                "summary": "Get orders for a specific table",
                "operationId": "ace94a7415231cc98cc05f0fe5ece963",
                "parameters": [
                    {
                        "name": "table_id",
                        "in": "path",
                        "description": "ID of the table",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders for the table."
                    },
                    "404": {
                        "description": "Table not found."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Category Controller",
            "description": "Endpoints for managing categories"
        },
        {
            "name": "Order Controller",
            "description": "Endpoints for managing orders"
        },
        {
            "name": "Product Controller",
            "description": "Endpoints for managing products"
        },
        {
            "name": "Review Controller",
            "description": "Endpoints for managing reviews"
        },
        {
            "name": "Table Controller",
            "description": "Endpoints for managing tables"
        }
    ]
}